+++ conjur/logic/credential_provider/credential_store_factory.py
@@ -26,14 +26,14 @@ class CredentialStoreFactory:
     """
 
     @classmethod
-    def create_credential_store(cls, use_netrc_flag: bool = None) -> CredentialsProviderInterface:
+    def create_credential_store(cls, force_netrc_flag: bool = None) -> CredentialsProviderInterface:
         """
         Factory method for determining which store to use
         """
         keyring_name = KeystoreWrapper.get_keyring_name()
-        use_netrc = use_netrc_flag or util_functions.get_netrc_flag_from_conjurrc()
+        use_netrc = force_netrc_flag or util_functions.get_netrc_path_from_conjurrc()
 
-        if keyring_name in SUPPORTED_BACKENDS and use_netrc is False:
+        if keyring_name in SUPPORTED_BACKENDS and use_netrc is None:
             if KeystoreWrapper.is_keyring_accessible():
                 return KeystoreCredentialsProvider()