+++ test/test_unit_utils.py
@@ -1,7 +1,7 @@
 import unittest
 
 from conjur_api.models import SslVerificationMode
-from conjur.constants import DEFAULT_CERTIFICATE_FILE
+from conjur.constants import DEFAULT_CERTIFICATE_FILE, DEFAULT_NETRC_FILE
 from conjur.util import util_functions as utils
 from conjur.data_object import ConjurrcData
 
@@ -30,14 +30,14 @@ class UserInputDataTest(unittest.TestCase):
         res = utils.get_ssl_verification_meta_data_from_conjurrc(True, data)
         self.assertEquals(res.mode, SslVerificationMode.CA_BUNDLE)
 
-    def test_get_netrc_flag_from_conjurrc_none(self):
+    def test_get_netrc_path_from_conjurrc_none(self):
         data = ConjurrcData(conjur_url="https://foo.com", account="some_account",
-                            cert_file="cert", use_netrc=None)
-        res = utils.get_netrc_flag_from_conjurrc(data)
-        self.assertEquals(res, False)
+                            cert_file="cert", netrc_path=None)
+        res = utils.get_netrc_path_from_conjurrc(data)
+        self.assertEquals(res, None)
 
-    def test_get_netrc_flag_from_conjurrc_true(self):
+    def test_get_netrc_path_from_conjurrc_true(self):
         data = ConjurrcData(conjur_url="https://foo.com", account="some_account",
-                            cert_file="cert", use_netrc=True)
-        res = utils.get_netrc_flag_from_conjurrc(data)
-        self.assertEquals(res, True)
+                            cert_file="cert", netrc_path=DEFAULT_NETRC_FILE)
+        res = utils.get_netrc_path_from_conjurrc(data)
+        self.assertEquals(res, DEFAULT_NETRC_FILE)